+++ src/LandingPage/AboutUs/AboutCard/AboutCard.js
@@ -4,11 +4,11 @@ import './AboutCard.css'
 
 export default function AboutCard({title, desc, icon}) {
   return (
-    <div className='AboutCard'>
+    <div className="AboutCard">
       <div className="icon">
         {icon}
       </div>
-      <div className='title'>{title}</div>
+      <h3 className="title separator-center">{title}</h3>
       <p>{desc}</p>
     </div>
   )
