+++ src/LandingPage/AboutUs/AboutUs.css
@@ -1,0 +1,11 @@
+.AboutUs {
+  display: grid;
+  grid-template-columns: 1fr;
+  padding: 30px 0;
+}
+
+@media (min-width: 600px) {
+  .AboutUs {
+    grid-template-columns: repeat(3, minmax(250px, 1fr));
+  }
+}
