+++ src/LandingPage/OurGoal/OurGoal.css
@@ -1,77 +1,39 @@
 .OurGoal {
-  text-align: center;
   --icon-dimensions: 30px;
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: 1rem;
 }
 
-.OurGoal h1 {
+@media (min-width: 600px) {
+  .OurGoal {
+    grid-template-columns: 1fr 1fr;
+    overflow: hidden;
+  }
+  .OurGoal .copy {
+    order: 1;
+  }
+}
+
+.OurGoal .iconAndText {
+  display: flex;
+  flex-direction: row;
+  margin: 1rem 0;
+}
+
+.OurGoal .slidingText {
   margin: 0;
-}
-
-.OurGoal .subtext {
-  font-size: 14px;
-  padding: 0 10px;
-}
-
-.OurGoal .container {
-  position: relative;
-}
-
-.OurGoal .overlay {
-  position: absolute;
-  height: 100%;
-  margin-left: 5px;
-  width: 200px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-}
-
-.OurGoal .icon {
-  display: flex;
-  justify-content: center;
-  color: lightgoldenrodyellow;
-}
-
-.OurGoal .iconAndText .slidingText {
-  color: lightgoldenrodyellow;
-  font-size: 14px;
+  padding-left: 15px;
   font-weight: bold;
-  margin: 1px;
-}
-
-.OurGoal button {
-  width: 60%;
-  border-radius: 5%;
-  padding: 10px;
-  margin: 10px auto;
-  box-shadow: 0px 0px 10px 2px rgba(0, 0, 0, 0.56);
 }
 
 @media (min-width: 600px) {
   .OurGoal {
     padding-bottom: 100px;
   }
-  .OurGoal h1 {
-    font-size: 50px;
-  }
-  .OurGoal .subtext {
-    font-size: 18px;
-  }
-  .OurGoal .overlay {
-    width: 50%;
-  }
-  .OurGoal .iconAndText .slidingText {
-    font-size: 18px;
-  }
+  
   .OurGoal .icon svg {
     width: var(--icon-dimensions);
     height: var(--icon-dimensions);
-  }
-  .OurGoal .quote {
-    font-size: 25px;
-    margin-top: 30px;
-  }
-  .OurGoal button {
-    font-size: 20px;
   }
 }
