--- src/PageView/AboutUs/AboutUs.css
@@ -1,10 +1,0 @@
-.AboutUs {
-  display: grid;
-  grid-template-columns: repeat(3, minmax(250px, 1fr));
-}
-
-@media (max-width: 720px) {
-  .AboutUs {
-    grid-template-columns: 1fr;
-  }
-}
